types: update Milestone to account for group milestones
Since upstream gitlab-ce commit v9.5.0-rc1~292^2 (Add group milestones
API endpoint, 2017-07-12) a Milestone may have either a `project_id`
or a `group_id`, making each optional.  Later we could consider making
this an `entity_id` with a custom `enum MilestoneEntityId` that can
contain either a `ProjectId` or a `GroupId` but this is sufficient
for now.