Support OAuth2 tokens for API access
OAuth2 tokens are supported in addition to personal access tokens.
The change is backwards compatible, the current behavior does not change
and existing code will keep using the personal access tokens.

A new set of constructors and a new builder method have been added
in order to specify that an OAuth2 token is being used.